% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chile}
\alias{chile}
\title{Chilean FSP (PM2.5) Data}
\format{
A data frame with 3,840 rows and 12 variables:
\describe{
  \item{date}{Date of the recording.}
  \item{month}{Month of the recording.}
  \item{year}{Year of the recording.}
  \item{name}{Name of the monitoring station.}
  \item{station}{Unique identifier of the monitoring station}
  \item{longitude}{Longitude of the station}
  \item{latitude}{Latitude of the station}
  \item{pm25}{FSP (PM2.5) in µg/m³}
  \item{hum}{Relative humidity in \%}
  \item{temp}{Temperature in °C}
  \item{windDir}{Wind direction in angular degrees}
  \item{windSpeed}{Wind speed in m/s}
}
}
\description{
This dataset contains concentration levels of fine suspended particulates,
commonly known as PM2.5, recorded across 64 Chilean monitoring stations 
between January 2020 to December 2024. The dataset also includes geographical 
coordinates for each station and additional covariates such as temperature,
humidity, wind direction, and wind speed. 
These data have been used in (Marin et al., 2026+).
}
\examples{
data("chile")
head(chile)

}
\references{
S. Marin, B. Long,and A. H. Westveld (2026+), Bayesian Nonparametric Modeling
of Dynamic Pollution Clusters through an Autoregressive Logistic-beta 
Stirling-gamma Process. \emph{arXiv}, {2AAA.BBBB}.
}
\keyword{datasets}
