% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering estimates}
\usage{
clustering(object, dist = "VI", return.clusters.means = TRUE)
}
\arguments{
\item{object}{An object of class \code{"dynclust"}.}

\item{dist}{A character string denoting which distance function should be 
used. Options are either \code{"VI"} for the  variation of information 
oss function or \code{"Binder"} for Binder's loss function. Default is 
\code{"VI"}}

\item{return.clusters.means}{Logical. Whether the means of each cluster 
should be returned. Default is \code{TRUE}.}
}
\value{
Clustering of the observations.
}
\description{
Identifies the partition that minimizes an user-specified distance function 
for an object of class \code{"dynclust"}.
}
\author{
Santiago Marin
}
