% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering_and_ari_matrices.R
\name{coclust_point}
\alias{coclust_point}
\title{Co-clustering matrix}
\usage{
coclust_point(S, time.point = 1L)
}
\arguments{
\item{S}{Matrix of cluster allocations.}

\item{time.point}{Integer indicating the time point at which we would like 
to compute the co-clustering probabilities at. Default is \code{1}.}
}
\value{
Co-clustering matrix
}
\description{
Computes the co-clustering matrix at a specific time point based on a matrix 
of cluster allocations.
}
\author{
Santiago Marin
}
