% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering_and_ari_matrices.R
\name{coclust_probs}
\alias{coclust_probs}
\title{Matrix of posterior co-clustering probabilities}
\usage{
coclust_probs(object, time.point = 1L)
}
\arguments{
\item{object}{An object of class \code{"dynclust"}.}

\item{time.point}{Integer indicating the time point at which we would like 
to compute the co-clustering probabilities at. Default is \code{1}.}
}
\value{
Matrix of posterior co-clustering probabilities
}
\description{
Computes the matrix of posterior co-clustering probabilities at a specific 
time point for an object of class \code{"dynclust"}.
}
\author{
Santiago Marin
}
