% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynclusts-package.R
\docType{package}
\name{dynclusts}
\alias{dynclusts-package}
\alias{dynclusts}
\title{dynclusts: Bayesian nonparametric modeling of dynamic clusters through an
autoregressive logistic-beta Stirling-gamma process}
\description{
The package allows users to identify latent dynamic clusters through an 
autoregressive logistic-beta Stirling-gamma (AR-LB-SG) process as described 
in Marin et al. (2026+).
}
\references{
S. Marin, B. Long,and A. H. Westveld (2026+), Bayesian nonparametric modeling
of dynamic pollution clusters through an autoregressive logistic-beta 
Stirling-gamma process. \emph{arXiv}, 2601.04625.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marinsantiago/dynclusts}
  \item Report bugs at \url{https://github.com/marinsantiago/dynclusts/issues}
}

}
\author{
\strong{Maintainer}: Santiago Marin \email{santiago.marinardila@anu.edu.au} (\href{https://orcid.org/0009-0006-0086-3721}{ORCID})

Other contributors:
\itemize{
  \item Bronwyn Loong \email{bronwyn.loong@anu.edu.au} (\href{https://orcid.org/0000-0002-1409-8892}{ORCID}) [contributor]
  \item Anton H. Westveld \email{anton.westveld@anu.edu.au} (\href{https://orcid.org/0000-0003-1671-0931}{ORCID}) [contributor]
}

}
