% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclustering_and_ari_matrices.R
\name{lagged_ARI}
\alias{lagged_ARI}
\title{Matrix of lagged ARI values}
\usage{
lagged_ARI(S)
}
\arguments{
\item{S}{Matrix of cluster allocations.}
}
\value{
Matrix of lagged ARI values
}
\description{
Computes the matrix of lagged ARI values based on a matrix of cluster 
allocations.
}
\author{
Santiago Marin
}
