% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynclusts_sim.R
\name{dynclusts_sim}
\alias{dynclusts_sim}
\title{Simulate dynamic clusters}
\usage{
dynclusts_sim(
  n.obs,
  time.points,
  n.clusts = 3L,
  n.covariates = 5L,
  country = "United Kingdom",
  tau2 = 2,
  rho2 = 1,
  varphi = 100,
  sigma2 = 1,
  pi.stay = 0.3,
  balanced = FALSE,
  spatial.effect = TRUE,
  covariates.effect = TRUE
)
}
\arguments{
\item{n.obs}{Number of locations}

\item{time.points}{Number of time points.}

\item{n.clusts}{Number of clusters. Default is \code{3}.}

\item{n.covariates}{Number of covariates. Default is \code{5}.}

\item{country}{Country to pick locations from. Default is "United Kingdom".}

\item{tau2}{Variance of the location-specific random effects. Default is 
\code{2}.}

\item{rho2}{Variance of the vector of coefficients, \eqn{\boldsymbol{\beta}}.
Default is \code{5}.}

\item{varphi}{Strength of the spatial correlation between two locations.
Default is \code{100}.}

\item{sigma2}{Sampling variance, assumed to be common across all clusters. 
Default is \code{1}.}

\item{pi.stay}{Probability that a given station will remain in the same 
cluster from one time point to another. Default is \code{0.3}.}

\item{balanced}{logical. Whether cluster sizes should be balanced or not.}

\item{spatial.effect}{logical. Whether spatial effects should be
incorporated.}

\item{covariates.effect}{logical. Whether covariate effects should be 
included.}
}
\value{
An object of S3 class, \code{"dynclust.sim"}, containing:
\itemize{
  \item \code{y}: A panel data matrix of size \eqn{T}-by-\eqn{n}, where each 
  of the \eqn{T} rows corresponds to a time point and each of the \eqn{n} 
  columns corresponds to location.
  \item \code{S}: A matrix of size \eqn{T}-by-\eqn{n} with the cluster
  allocations.
  \item \code{coords}: A two column matrix with the geographical coordinates 
  of each location. The first column corresponds to the longitude, while the
  second column corresponds to the latitude.
  \item \code{theta}: A matrix of size \eqn{T}-by-\eqn{n} with each of the
  \eqn{\theta_{ti}} parameters.
  \item \code{sigma2}: A matrix of size \eqn{T}-by-\eqn{n} with each of the
  \eqn{\sigma_{ti}^{2}} parameters.
  \item \code{gamma}: If \code{spatial.effect = TRUE}, a vector of size 
  \eqn{n} with the location-specific random effects, \eqn{\gamma_{i}}.
  \item \code{K}: If \code{spatial.effect = TRUE}, the squared exponential
  kernel.
  \item \code{beta}: If \code{covariates.effect = TRUE}, a vector of size
  \code{n.covariates} with the coefficients associated with each of 
  covariate.
  \item \code{X}: If \code{covariates.effect = TRUE}, a list of size \eqn{n},
  containing matrices of size \eqn{T}-by-\eqn{p} of covariates associated
  with each location. More precisely, each of the \eqn{T} rows contains 
  \eqn{p} predictors at time \eqn{t\in\{1,\dots,T\}} associated with the 
  location \eqn{i\in\{1,\dots,n\}}.
}
}
\description{
This function simulates dynamic clusters as in Marin et al. (2026+).
}
\references{
S. Marin, B. Long,and A. H. Westveld (2026+), Bayesian nonparametric modeling
of dynamic pollution clusters through an autoregressive logistic-beta 
Stirling-gamma process. \emph{arXiv}, 2601.04625.
}
\author{
Santiago Marin
}
